reduced concrete cost
TBH sand is not a common as you think calc. I had to rape a desert to
rebuild my base with concrete. Also if its ok with you i want to try to
make an item called concrete mix. This will replace parts of the recipes
and maybe when right clicked on water will make concrete hint hint