Worked on how sentries attack targets
Change ammo item to return a Damage Source and damage amount rather than
cause the damage. As well change the turret platfrom to call a method in
the ammo item to consume the items. This should allow use later to
create ammo belts, and damage-able ammo items.

More work is still need to improve and use these changes to the fullest.