Made Sentries take damage
Still working on the power issue though i think its still only with
buildcraft as mekanism power works.

As well sentries do take damage and this has been tested. I used a fake
entity that is spawned at the tile location and is a little large than a
block in size. The entity keeps getting push out of the block so some
shots will missing. However, most will hit as the entity is reset as
fast as possible. I will try to see what is causing the position reset
as its not in the Entity.class.