Working on fixing railgun
Several issues with the rail gun are present and baddly need fixed. So
while i'm fixing these issue i'm going to overhaul the rail gun and
improve the base sentry class file. Planned improvement are too have
sent shot to client call render shot client side. This should fix the
render issues with the railgun. As well going to re-do the packet system
to handle packets better for the sentries.