Toyed with missile renderer
Don't worry i reverted the changes since they didn't work. However, i
did noticed an improvement in performance from my efforts. Though since
the outcome was not right i had to scrape it. I think though that VBO
might fix this issue but i have to learn how to use it. Not even sure if
minecraft supports this kind or rendering but it shows promise. As well
we can convert to obj models to help out.