A lot of work on module setup
I noticed that this system can be used for just about anything. So don't
be surprised if it starts to show up in a few of my mods. As well this
is only 3% finished as i need to lay in a lot more ground work. However,
before you say its too complex the user will not noticed an over all
level of complexity. The code however will be a bit higher up in
complexity to cover all the cases of its use. The player will just treat
it as a crafting system for missiles. In the same way that flans plans
work but more world based, less crafting table. This system i  hope will
be very scalable that way it can be expanded at ease. As well players
can have fun finding that perfect mix of power, and technology in their
missiles. We also need to work on letting other mod use this system to
add effect to your missiles.