Started doing research on blast physics
The idea is to improve the current way we do our blast to both react
better to the world. As well reduce to run time the explosion by pre
calculating the blast then carrying out the block damage to the world.
As well we want to create a more realistic blast so that blast shielding
actually plays an important rule to base designs.