- Fixed Radar peripheral for ComputerCraft methods always returning "insufficient electricity!"
Essentially, the problem was that the Radar would be given energy, would use the energy and update, and then ComputerCraft would call a method, before more power was given to the Radar.
Now we just mark it as being powered/unpowered during the tick, and the ComputerCraft call will just check that.