- Bug Fix: Client should no longer randomly crash when opening an ICBM GUI while having an updating item in inventory.
This happens because the ICBM GUI screens does not have an inventory, so when the client opens a GUI window locally before the server, and the server tries to update an item in the (now non-existent) player inventory, it crashes.
Now we delay opening the GUI until the server tells us to, but a better fix would probably be to just enable the player inventory in the ICBM GUI's.