Fixed desync causing door tile to alterate
Server and client were fighting on who had the right tile. This was mainly an issue with the tile only being set client side. While the TileMulti was still sending desc() packets to the client. When it did so it would reset the render shortly. To fix this the sync check is now run on both sides often.